Blskey migration

+++ pipeline/go.sh
@@ -341,7 +341,7 @@ EOT
 EOT
       curl -X POST https://${configs[explorer.name]}:${configs[explorer.port]}/reset -H 'content-type: application/json' -d@explorer.reset.json
    fi
-   mv -f explorer.reset.json logs/$TS
+   [ -e explorer.reset.json ] && mv -f explorer.reset.json logs/$TS
 }
 
 function do_wallet_ini
